Implemented support for SASL_SSL parameters.
Some debugging log statements are included. We can revise these as the
code is refined. This is intended to fulfil the basic requirements of
https://github.com/commercetest/pepper-box/issues/1

There's little or no error checking or reporting. This might be useful
to add generally i.e. not limited to the recent additions as currently
neither the jmeter GUI nor the jmeter log are particularly insightful
when there are problems with the values assigned.