Added a simple way to create per jmeter-thread topic names.
The name consists of the user-specified name with a suffix of
.{thread-number} e.g. mytopic.0 mytopic.1 etc The names are written to
the jmeter log file to make them a bit easier to locate. If and when we
create a consumer to pair with the producer we'll want to read from the
same Kafka topic the respective thread wrote to. As we don't yet have
that capability it's not necessary to store the topic name & perhaps the
consumer would be able to calculate it anyway...

This feature is controlled by a YES/NO value property:
  GENERATE_PER_THREAD_TOPICS = "generate.per-thread.topics";

The default value is NO as this enables the current tests to run without
complaint.

This is intended to implement
https://github.com/commercetest/pepper-box/issues/5