Fixed the SASL JASS Config bug.
We were passing in hardcoded example values rather than the ones
provided in the configuration parameters (in user.properties). Also
there was a typo in the java package path on the same line. The jmeter
GUI form made this bug particularly hard to spot as the text is
truncated, and testing this feature requires a Kafka instance configured
with SASL_SSL so the bug hid in plain sight.