Added consumerGap measurement and an optional prefix to topics, etc.
Tidied up the functionality and it now calculated the time taken from
when the message was generated and when it was recieved.

Occasionally the time is negative or zero when the tests are run against
localhost Kafka. Something's not quite accurate/correct.

The optional prefix is to facilitate some specific testing I'm doing.
You're unlikely to need it.
 consumerGap duration