Created a groovy script that polls whether a topic exists on zookeeper.
This will enable us to measure how long it takes for the topic to be
created, and also to only start trying to consume messages once it does.

This is a proof-of-concept, wrapped in a small test plan to make it easy
to run for now. We will refine the code and then integrate it with the
main codebase and examples.