The consumer script now waits until the topic exists before consuming messages.
The code is rough and imbedded in the midst of the current script. TODO:
tidy up the code. I wanted to implement this logic in a pre-processor
step but haven't yet found a good way to stop the main sampler (the
Consumer) from running if we don't find the topic in time.