Added multi-topic support to the Load Generator to mirror the feature in the jmeter code.
Also revised and extended the automated tests for the Load Generator. These tests
indicate the basic concept works and the load generator can generate many messages.

However my testing indicates that the load generator takes a while to get started
and needs to be allowed sufficient time (around 7 to 8 seconds on my macbook 12") to
start generating messages in earnest. So there's plenty of scope to improve the testing
and analysis of the load generator. Also, for users of the generator, configure the tests
to run for long-enough (say a minimum of 60 seconds) if you want to actually generate load
in earnest.